Add functionality for operators and equals
Pressing an operator (+,-,*,/) saves the current display value to either
num1 or num2 by checking if num1 is a number. Clicking equals also takes
the display value and assigns it to either variable.

This works for a single calculation, however if attempting to chain
calculations the logic is flawed as num1/2 are assigned improperly.
Thus it is the next step of development: cleaning up function logic.