Create logic for equals and operators
Equals now unassigns the operator variable so the script doesn't
eroneously keep doing additional math not being inputted by the user.

The script also keeps track of a new state: if the program recently
assigned a value to num1 or num2, it will know the maintain the number
that was saved on the display until a new number is being selected.
That way the user can see the number they've inputted in between
operators and can see the result of an equals sequence in between chains.