Add logic for storing numbers and operating them
Pressing an operator stores the number in the display to firstNum and
the operator to its own variable. Pressing equals stores the display
number to secondNum, calls operate() and puts it into the display.

This doesn't account for many things still, such as if they do more than
one operation at a time, but this is for getting the most basic
functionality in there and continue to add complexity from there.