Add logic to prevent prematurely clearing display
displayingResult boolean variable was put into place to let the program
know the difference between a user-inputted number and an operate()
result number. Essentially, if it's the latter, user-input numbers
should clear the display first before proceeding. Otherwise, just
concatenate the string as normal.

Before this, every time executeEquals() was called, it wouldn't be able
to display the result as it gets cleared during the operaterBtn event.
This is prevented by an "on/off switch" as described above.